final lab2

+++ kernel/proc.h
@@ -93,7 +93,7 @@ struct proc {
   int killed;                  // If non-zero, have been killed
   int xstate;                  // Exit status to be returned to parent's wait
   int pid;                     // Process ID
-  int mask;                    // Mask for trace
+  uint64 mask;                    // Mask for trace
 
   // these are private to the process, so p->lock need not be held.
   uint64 kstack;               // Virtual address of kernel stack
